aeval/orchestrator/requirements.txt
fastapi>=0.115
uvicorn>=0.34
psycopg2-binary>=2.9
rq>=2.1
redis>=5.0
pydantic-settings>=2.0
httpx>=0.27
